backend-service/requirements.txt
opencv-python==4.8.0.74
fastapi==0.103.1
uvicorn==0.23.2
pillow==10.0.0
numpy==1.25.2
python-multipart==0.0.6
sam2
torch>=2.0.0
psutil>=5.9.0
scipy>=1.11.0
requests>=2.28.0
